classifier/requirements.txt
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
xgboost>=1.7.0
pandas>=1.5.0
numpy>=1.21.0
regex>=2022.0.0
PyYAML>=6.0
joblib>=1.2.0
